[CHORE] Date types  (#88)

+++ apps/backend/src/database/migrations/1677775665682-migration.ts
@@ -10,7 +10,7 @@ export class migration1677775665682 implements MigrationInterface {
 
   public async down(queryRunner: QueryRunner): Promise<void> {
     await queryRunner.query(`
-        DELETE FROM PAYMENT WHERE AMOUNT = 0::numeric;
+        DELETE FROM PAYMENT WHERE AMOUNT = 0;
     `);
   }
 }
